feat: Responsive design and offline availability feature added.
The site was made more responsive. All of the page elements are usable and visible in any viewport, including desktop, tablet, and mobile displays. Images are not overlapping, and page elements are wrap when the viewport is too small to display them side by side.

Also the Page can be visited offline.